Add pyproject.toml, remove setup.py
