--- fvhiot/__version__.py
@@ -1,3 +1,0 @@
-VERSION = (0, 0, 1)
-
-__version__ = '.'.join(map(str, VERSION))
