+++ pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = [
+    "setuptools>=42",
+    "wheel"
+]
+build-backend = "setuptools.build_meta"
+
+[project]
+dependencies = [
+    'Flask',
+    'starlette',
+    'certifi',
+    'kafka-python',
+    'msgpack',
+]
