+++ setup.cfg
@@ -1,0 +1,23 @@
+[metadata]
+name = fvhiot
+version = 0.1.0
+author = Aapo Rista
+author_email = aapris@gmail.com
+description = Tools for Forum Virium Helsinki's IoT data pipe processes
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/ForumVirium/FVHIoT-Python
+project_urls =
+    Bug Tracker = https://github.com/ForumViriumHelsinki/FVHIoT-python/issues
+classifiers =
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+packages = find:
+python_requires = >=3.7
